ref: break down cookieListing logic into a seperate hook
